pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-agents-hub"
version = "0.1.0"
description = "OpenAI-compatible multi-agent router/supervisor with markdown memory."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "pydantic>=2.8.0",
  "pyyaml>=6.0.1",
  "python-dotenv>=1.0.1",
  "litellm>=1.50.0",
  "duckduckgo-search>=7.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
]

[project.scripts]
ai-agents-hub = "ai_agents_hub.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
